rename mathstuff to linalg
