+++ src/ellipsoid.c
@@ -6,7 +6,7 @@
 #include <SDL2/SDL.h>
 #include "camera.h"
 #include "log.h"
-#include "mathstuff.h"
+#include "linalg.h"
 #include "rect3.h"
 
 static bool ellipsoid_intersects_plane(const struct Ellipsoid *el, struct Plane pl)
