+++ src/linalg.h
